[#166172190] Replace the seeds file with the one generated by the gem
We've added a seeds file template to the dahlia_data_models gem, so now running rails g dahlia_data_models:install will copy over that seeds file into the dahlia_admin app. We'll be maintaining the seeds file in the gem from now on, and only updating the seeds file in this repo with the generated version from the gem.